#3 changed type of price from int to double

+++ src/model/IArticle.java
@@ -5,10 +5,10 @@ public interface IArticle {
 	public String getDescription();
 	public void setDescription(String description);
 	
-	public int getPrice();
-	public void setPrice(int price);
+	public double getPrice();
+	public void setPrice(double price);
 	
 	public int getYear();
-	public void setYear(int year);
+	public void setYear(int year);	
 
 }
